(PUP-11518) Fix type checking with deferred functions
Using a deferred function for a typed parameter value of a class
results in a compilation error stating that the type defined for
the parameter does not match the “Deferred” type at compilation
time. This error prevents the use of Deferred functions for the
values of any typed class parameters.

This change updates the type check to introspect Deferred
functions for the return_value type of that Deferred function.