(PUP-11522) Ensure test is using the SSL_CERT_FILE
Previously, the test passed the CA cert to `create_context` and set
SSL_CERT_FILE, so it was possible for the test to pass when it shouldn't. Now
only do the latter to verify we can indeed connect when loading a CA from the
"system store" aka wherever openssl looks for trusted certs.

Also rename "system store" variable to "cert_file" to avoid confusion later when
the file is used to contain certs in an "external" store.