(PUP-11522) Include client cert if requested
It's now possible to create a `system` context that performs mutual TLS and
trusts certs in the system store and/or the external store, based on
`Puppet[:ssl_trust_store]`.

If the agent doesn't have a client cert yet, then it will be ignored. For
example, when running `puppet apply` without a client certificate and trying to
apply a file resource with an "https" source.