(PUP-9998) Check agent disabled status after sleeping
Puppet's disabled status checking was subject to race conditions. It checked the
status once at the beginning of its run, but between the time it checked and
when the agent woke up from splay, someone could disable the agent.

This commit adds additional checks in situations where we may have slept for
arbitrary amounts of time.

Note when running onetime, the `Puppet::Agent` class handles splaying. Otherwise,
splay is handled by the job scheduler, before `Puppet::Agent#run` is called and
the job scheduler passes `splay: false` as an argument to the `run` method.