(PUP-11540) Update task_info_service to return private/description
Updates the task_info_service implementation to return all task metadata
for each task when tasks_for_environment is called. This requires
reading metadata json files for each task.

Testing was done to ensure the additional file reads for metadata json
files would not impact performance of puppetserver. In tests with an
environment with > 3000 tasks, the difference in total time to process
all task data was only any increase of 0.5 - 1 second..