(PUP-11629) Handles empty task metadata file
An empty string is valid JSON per the IETF standard, but a bug in
MultiJSON causes it to throw MultiJson::ParseError when loading
an empty JSON file.

This commit updates tasks to check for an empty JSON file to work
around this issue and instead will return an empty hash.