(PUP-11662) Allow legacy facts to be excluded
Adds a puppet setting `include_legacy_facts`. If set to false, then legacy facts
will not be sent to puppetserver and will not be available during `puppet apply`
compilation.

This commit doesn't cause legacy facts to be blocked, so they are still
available via `Facter.value(:osfamily)` for example. This is important to
maintain compatibility with `defaultfor`, `confine`, etc statements used to
select the most suitable provider.