(PUP-11654) Remove keys from resource_statuses when omitting excess data from report
Previously, when :exclude_unchanged_resources was set to true, the
key name would remain in the `resource_statuses` map with an
empty hash. This ended up breaking the puppetdb report class
because of an assumption that elements in `resource_statuses`
would be a Puppet::Resource::Status instance. This change removes
the key entirely from `resource_statuses`, so puppetdb no longer
errors when interating of the keys.

Additionally, there is a nebulous definition of what really a
`unchanged_resource` might be; this change expands to to be
any resource that is not failed, failed_to_restart, out_of_sync,
or skipped. In the future, it may be a good idea to add a
state named 'unchanged' instead of relying on the absence of
those other 4 attributes listed.