(PUP-10688) Allow tidy to pass on metaparameters
Prior to this commit, if a tidy resource contained a metaparamter
like "schedule," that metaparameter was not passed on to generated
child file resources.

This commit uses the copy_metaparams method to allow tidy to generate
child file resources with the same metaparameters (except for "alias").

It also updates and simplifies some older spec tests.