(maint) Pin RuboCop gem
This commit pins the RuboCop gem in the Gemfile to version 1.28.0.
This version was chosen because it's the last version of RuboCop
that supports Ruby 2.5, which we still use to run some RSpec tests
for Puppet.

Additionally, this commit regenerates the .rubocop_todo.yml file
using the updated gem version.