(PUP-9648) Use binary_file instead of file
The `file` function returns a `String`, not `Binary`, so if the string contains
data that is not valid UTF-8, then puppetserver will fail to serialize the
catalog as `rich_data_json`. However as described in PUP-10928, puppet will
fallback to either `rich_data_msgpack` (if msgpack is installed as a
puppetserver gem) or fallback to `pson`.

Rich data support was added to agents in Puppet 6, so just use the `binary_file`
function instead of relying on the fallback behavior.