(PUP-11787) Warn if we fail to serialize and fallback
If puppetserver falls back to PSON when serializing a catalog, then conversion
of rich data will be lossy, but the problem won't be noticed until the catalog
is applied on the agent.

This converts the debug message to a warning to assist troubleshooting. This
affects both `find` and `search` operations for all indirected types.