(PUP-10928) Add allow_pson_serialization setting
Currently, Puppet will fall back to PSON when unable to serialize to JSON. This
can cause issues since rich data types cannot be serialized via PSON. Due to
this, PSON will eventually be depreceated and removed in future releases. This
commit adds a setting, :allow_pson_serialization, which defaults to false in
Puppet 7 and will default to true in Puppet 8. When set to false, a warning
will be raised when falling back to PSON and when set to true, an error will be
raised instead. This option affects both puppetserver's configuration management
service responses and when the agent saves its cached catalog.