(PUP-11348) Remove dead test
The test was confined to only run on Ruby < 3.0, but the minimum required
version is 3.1, so delete the test.

Ruby 3+ correctly uses wide Win32 APIs to access environment variables and
transform them between UTF-16LE and UTF-8 regardless of the current code page.
So it is possible to get/set variables whose unicode code points don't map to
the current code page and that is already tested in util_spec.rb.