(PUP-10928) Fix indirected_routes_spec.rb for puppet#main
Unlike puppet#7.x (where the :allow_pson_serialization setting was first
added), by default in puppet#main, pson is not a supported format so the rspec
tests for :allow_pson_serialization were not able to test the desired behavior.
This commit stubs out Puppet.features.pson? in those tests to stimulate that
the puppet-pson library is present so pson is a supported format returned by
accepted_response_formatters.