(PUP-10639) Add ca_refresh_interval setting
It defaults to once per day, because if the CA expires, the entire
infrastructure is dead until the CA is updated on the agent. And the
puppetserver load to stat the `ca_crt.pem` is negligible as compared to the load
caused by pluginsync (each agent run causes puppetserver to checksum every file
in the environment's modulepath).