(PUP-10639) Refresh CA certs
Refresh CA certs if the last update time plus the `:ca_refresh_interval`
is in the past.

If the certs are updated, then force the CRLs to be updated too, since if the CA
bundle includes a new root or intermediate CA, we'll need its corresponding
non-expired CRL, to check the full chain.

If the certs aren't updated or there's an networking error, then continue using
the CA started with, and try again based on the next `:ca_refresh_interval`.