(maint) Removes AIX 5.3-specific setting
bb86ff6 added a conditional to Puppet's default settings to account for
unique behavior for diff on AIX 5.3, which has long been end-of-life
and is no longer supported by Puppet or Puppet Enterprise.

This commit updates Puppet's default settings for diff and associated
spec tests to remove that AIX 5.3-specific conditional.