(PUP-11869) Retry failed CA/CRL refreshes sooner
Prior to this commit, any time a node attempted to refresh its
certificate authority (CA) or certificate revocation list (CRL),
regardless of whether it was successful in doing so, Puppet would update
the modified timestamp (mtime) on the CA or CRL file.

Puppet uses the mtime to decide when to attempt to update a CA or CRL.
As a result, if Puppet attempted but failed to updates either of those
file, it would need to wait the full refresh interval before trying
again.

This commit changes Puppet to only update the modified timestamp on the
CA or CRL files if they have been successfully refreshed. This means
that if Puppet fails to update its CA or CRL, it will try again the next
run interval instead of waiting for the full CA or CRL refresh interval
(which should be greater than the run interval).