(maint) Simplifies SSLContext intialization
Ruby 2.5 added the `keyword_init: true` option to Struct.new to enable
keyword arguments.

This commit updates the Puppet::SSL::SSLContext to take advantage of
this language feature and removes a workaround needed for older versions
of Ruby.

Additionally, Ruby 3.2 addeed the ability to initialize a Struct with
keyword arguments without the `keyword_init: true` option. This commit
adds a comment noting that, so it can be removed in the future when
Puppet drops support for Ruby < 3.2.