(PUP-11897) Handle EOF in order to avoid busy-loop
When EOF is reached, calls to IO.select() will return immediately,
resulting in puppet executing the loop without any internal blocking.
This causes up to 100% usage of a CPU core until the child finishes.

If the child does its own stdout/stderr redirection before a long
operation, for example, this can result in much wasted CPU usage.

There is no need to execute the body of the loop after EOF is reached,
so switch the Process.waitpid2 flags to block.