(PUP-11846) Handle Deferred(Sensitive) file content
If file `content` is assigned a Sensitive value, then normally the `content` and
`ensure` properties are marked as sensitive early on when the
`Puppet::Type::File` instance is created[1].

The same is true if a sensitive value is wrapped in deferred and deferred
functions are preprocessed, since that happens before the `Puppet::Type::File`
instance is created.

However, if a sensitive value is wrapped in deferred and deferred functions are
not preprocessed, instead evaluated lazily as each resource is evaluated, then
the `munge` method may receive a `PSensitiveType::Sensitive` value. So we must
unwrap the file content to compute the should/desired checksum and we must mark
the `content` property as sensitive so that the output is redacted:

    Notice: /Stage[main]/Main/File[/tmp/test.txt]/content: changed [redacted] to [redacted]

Also puppet handles created vs modified files differently. In the latter case,
only the `ensure` property is managed[2], so we have to mark that as sensitive
too, resulting in:

    Notice: /Stage[main]/Main/File[/tmp/test.txt]/ensure: changed [redacted] to [redacted]

[1] https://github.com/puppetlabs/puppet/blob/fde57136980d63bd2a64532ad74f750cc3b73bc7/lib/puppet/type/file.rb#L977-L982
[2] https://github.com/puppetlabs/puppet/blob/fde57136980d63bd2a64532ad74f750cc3b73bc7/lib/puppet/transaction/resource_harness.rb#L81-L92