Lint/Void
Return values from `initialize` are ignored, unless you override `new` and change
the behavior. We do override `new` in a few places, but none of the ones I
changed do.

Ruby returns the argument passed to a setter method instead of what the setter
returns, so there's no point in pretending.