Lint/ConstantDefinitionInBlock (faces)
Constants defined in blocks aren't namespaced and instead are created in the
top-level namespace, e.g. DEFAULT_SECTION.

Define a new module for each face in which to store their respective constants
and deprecate the old top-level constants. A deprecation warning will only be
generated if the legacy constant is accessed when running ruby with -W2.