Lint/ConstantDefinitionInBlock (metaprogramming)
When metaprogramming types, parameters and providers, we defined constants
within the block, such as

    Puppet::Type.type(:package).provide(:xxx) do
      SOMETHING = ...

The constant isn't namespaced in the way you'd expect and instead is defined at
top-level, potentially clashing with non-Puppet code. In order to define
constants on the generated class:

   Puppet::Type::Package::ProviderXXX

use `const_set` instead. But continue providing the "legacy" constant to
maintain backwards compatibility and mark it as deprecated. A deprecation
warning will only be generated if the legacy constant is accessed when running
ruby with -W2:

    $ bundle exec ruby -W2 -rpuppet -e "Puppet::Type.type(:package).provider(:yum); RPM_VERSION"
    ...
    -e:1: warning: constant ::RPM_VERSION is deprecated