(PUP-11841) Fix encoding of empty String
In 2077971499d3cfc3e33b6eba4c91f43f82f1214c empty string literals where
replaced by a call to `String.new` in preparation for moving to
frozen/immutable strings.

However, as stated in the String#new documentation, when no value is
passed to `#new` the returned string has the 'ASCII-8BIT' encoding
instead of the default one (which is assumed to be 'UTF-8' by Puppet).

This cause regressions in some areas of the code, for example when using
the concat module with exported resource built using epp templates, the
incorrect encoding cause the fragment to be misinterpreted and is base64
encoded.  When collected, the base64 representation of the string is
used instead of the actual value of the string, as reported here:
https://github.com/voxpupuli/puppet-bacula/issues/189

Replace calls to `String.new` with `''.dup` which use the current
encoding.  Do not change the few explicit but redundant occurrences of
`String.new.force_encoding('ASCII-8BIT')` (so that the intent is clearly
visible).  Where appropriate, slightly adjust the code for better
readability.