(PUP-11928) Give more specific error messages for /puppet/v3/tasks API
Currently, when a task is omitted due to a failure, Puppet Server logs an
exception with a generic error message and provides no details on how to
resolve the error: `ERROR [qtp1527520408-905] [puppetserver] Puppet Failed to
validate task`

This commit removes the generic message that overrides the original error message in
the rescued error. Instead, the rescued error message is logged.