(PUP-11947) Pin FFI to 1.15.5
FFI >= 1.16.0 breaks FFI modules in Facter, which is a dependency of
Puppet. While the latest release of Facter (4.5.0) pins FFI to 1.15.5,
Puppet's gemspec adds a Windows-specific runtime dependency on FFI set
to > 1.9.24 < 2, which currently resolves to the latest version of FFI
(1.16.2).

This commit pins FFI to 1.15.5 to address the incompatibility between
Facter and FFI >=1.16.0.