(PUP-11974) Only warn when Ruby's verbose is enabled
In ruby 2.7.x, the rb_file_exists_p and rb_dir_exists_p methods call
rb_warning to log that the methods are deprecated[1][2] However, the
rb_warning method is a noop if $VERBOSE is nil or false[3] To preserve
the same behavior, only warn when $VERBOSE is truthy. Also include the
class name and path in the warning so we can identify and fix the issue.

[1] https://github.com/ruby/ruby/blob/1f4d4558484b370999954f3ede7e3aa3a3a01ef3/file.c#L1819
[2] https://github.com/ruby/ruby/blob/1f4d4558484b370999954f3ede7e3aa3a3a01ef3/dir.c#L3301
[3] https://github.com/ruby/ruby/blob/1f4d4558484b370999954f3ede7e3aa3a3a01ef3/error.c#L336-L338