Use resource titles as identifiers in preloading
When using composite namevars for resources then `resource.name` always
`nil`. Prior to this it was used as a key in a hash when prefetching
resources. By using the title it should work, since that's guaranteed to
be set and unique.

This can break providers where the name is different from the title.