Make ParsedFile provider work with composite namevars
Previously it was assumed that the name was the key. This changes the
implementation to instead match on the key attributes. This allows it to
work with composite namevars. While there was previously code for
multiple target support, this makes it possible to use the same key in
multiple files.

It also gets rid of some code that wasn't needed.