(maint) Only resolve environment dirs if versioned dirs are enabled
The previous implementation would not set the resolved_path variable of
an environment to the actual resolved path unless the
"report_configured_environment" was set to false. This should not matter
since the externalize_path method will check the
"report_configured_environment" setting.

The implementation should probably either always set the resolved_path
to the resolved path, or only set it if "versioned_environment_dirs" is
true. This patch sets it only if the "versioned_environment_dirs" setting
is true.