(PUP-11685) Update total fact count
Puppet calculates the total number of facts on an agent node to warn
the user if it will cause performance issues with PuppetDB. PuppetDB
performance can be impacted when the leaf count of a factset is too
high.

Prior to this commit, Puppet inconsistently calculated facts when
generating a warning. This did not count individual array elements; for
example, the processor.models fact with a fact per processor core would
only be counted once. Additionally, hash keys were incorrectly each
counted as individual facts when they would not represent separate leaf
counts in PuppetDB.

These behaviors lead to a slight discrepancy between what Puppet would
report to the user and what the total fact count ended up being in
PuppetDB.

This commit updates Puppet to more accurately count facts to reflect
a factset's leaf count in PuppetDB.