(PUP-11937) Skip autorequire if exec's param is deferred and lazily evaluated
If an exec's 'command' parameter is deferred and lazily evaluated, then we can't
autorequire the file that the command will execute. This is because we won't
know what the command will be until the exec is evaluated. But that means it's
already too late to add a relationship from the file resource to the exec
resource. In this situation an explicit relationship must be added to ensure
the file resource is evaluated before the exec. The same is true for the
'onlyif' and 'unless' exec parameters.