Lint/RescueException
Rescuing Exception is generally not recommended, because it will also rescue
SystemExit and SystemStackError (and more), preventing an application from
gracefully exiting.

Rescuing Exception is sometimes needed when calling 3rd party code, because that
code might raise an exception that doesn't inherit from StandardError. For
example, autoloading a provider might raise a LoadError if the provider relies
on a missing gem. Since LoadError extends ScriptError which extends Exception,
it is not caught when doing `rescue => e` since that only rescues StandardErrors.

In cases where we rescue Exception and immediately re-raise, then ignore the
cop, since a SystemExit will cause us to exit.

If we rescue the Exception, but try to continue along, then first rescue
SystemExit and re-raise that.

The Windows daemon code is fragile, so just ignore the cop there.