Remove redundant wrapper method
The is_constant_defined? method was added in 3d43d86 to account for
behavior differences between Ruby 1.8 and 1.9. It was simplified in
84a9cb8 as part of removal Ruby 1.8 support and the method exists now as
a fairly transparent wrapper of Ruby's built-in Module#const_defined?
method.

is_constant_defined? has been marked as private since 79889af in 2013.

This method removes the is_constant_defined? method and substitutes it
with Module#const_defined? instead.