Improve variable assignment readability
The assignment of the @type instance variable in
Puppet::Pops::Context#register_ptype resulted in poor readability due to
the number of arguments in the Pcore::create_object_type method call and
the number of those arguments that are nested hashes.

This commit reformats this assignment for easier readability and shorter
line lengths.