Copy the mtime on puppet generate types
When installing a new module it can maintain the file timestamps from
the original tarball (like g10 does). puppet generate types uses the
timestamp of when it ran.

If you update a module to a release that was created before it ran
puppet types generate, it will never update the types. This can happen
with a module downgrade, or just when you regenerated types manually. By
using --force the timestamps are ignored, but that's only a workaround.

This new approach copies the mtime from the source file to the generated
cache file. It is only considered up to date if the mtime matches.