Correct Red Hat service providers
SystemD has been the default init system for Red Hat Enterprise Linux
and its derivatives since RHEL 7 was released in 2014. In 928cf67, we
attempted to reflect this in Puppet's service providers.

However, this change did not have the desired effect as the RedHat
service provider was still specified as the default for systems with the
os.family fact of "redhat." Tests relating to services started failing
after 928cf67.

I do not believe that this change will have any undesired effects on Red
Hat derivatives like Alma or Rocky because those distributions also
use SystemD as their init system and use the same versioning scheme as
upstream.