Mock Debian OS in SystemD test
In e660355, we added an additional test for nonexistent services,
including a test for a Debian-specific code path that runs invoke-rc.d.

Because this code path is only invoked on Debian systems, this test
would regularly fail for any other platforms.

This commit mocks the os.family fact for the nonexistent services tests.