Add rubocop-performance extension
This commit adds the rubocop-performance extension to Puppet as part of
its Gemfile and .rubocop.yml configuration file.

I've specified version 1.13.3 as it's the last version compatible with
Ruby 2.5, which mirrors the version of Rubocop we have pinned and the
version of Ruby that is targeted in the Rubocop configuration file.

This commit also regenerates the Rubocop todo file and manually adds
pending Performance cops to both the configuration and todo files.