Add rubocop-rspec extension
This commit adds the rubocop-rspec extension to Puppet's Gemfile and
Rubocop configuration file, pinning it at 2.10.0 as it's the last
version to support Ruby 2.5 (which the rest of our Rubocop
configuration is set up to support).

This work includes explicitly adding all pending RSpec cops to the
Rubocop config file and enabling them.