Update Rubocop gems and configuration
Prior to this commit, Rubocop was configured to use Ruby 2.5 syntax and
all Rubocop-related gems were pinned to the last version to support Ruby
2.5. This was carried over from the Puppet 7 series, which specifies
  Ruby 2.5 as its minimum Ruby version.

This commit sets Rubocop's TargetRubyVersion to 3.1, the same as the
minimum Ruby version set in Puppet 8's gemspec. Additionally, this
commit relaxes all Rubocop-related gem version specifiers in the Gemfile
to allow for newer versions of the gems.