Resign modified client cert
These tests modify the not_before & not_after times for the client cert
and verify the cert is rejected because the cert is "not yet valid" and
"expired" respectively.

It just so happened openssl < 3.2, checks the times before checking the
signature. But openssl 3.2, checks the signature first. So regenerate the
signature using the intermediate CA's private key.