Remove Accept-Encoding header on redirect
Prior to this commit, Puppet would copy all request headers in an HTTP
redirect, including Accept-Encoding. In some cases when HTTP compression
was enabled, the response would fail to get decompressed, then would
fail to get parsed and trigger a vague error.

This commit strips the Accept-Encoding headers on redirect, allowing
Ruby's built-in Net::HTTP to both compress and decompress the traffic.