(PUP-11993) Move Style/PreferredHashMethods to .rubocop.yml
This commit moves Style/PreferredHashMethods from .rubocop_todo.yml
to .rubocop.yml to indefinitely disable the cop. This is because enabling
this cop causes 11 failures in rb_tree_map_spec.rb that relate to core
functions in trees, like deleting a node and returning nil when a key
can't be found.