(PUP-10214) Avoid replacing dangling symlinks on 'replace => false
Puppet::FileSystem.exist? explicitly follows symlinks, so the check for
in-sync will fail on a dangling symlink. A dangling symlink is still an
existing file, however, so when 'replace => false' is in effect, puppet
should not replace it.

Change the test to use Puppet::FileSystem.exist_nofollow? (which
ultimately uses File.lstat)  in order to avoid following symlinks.