Calling "pkg update -n" on Solaris too frequently causes problems with zones
On a Solaris system, running "pkg update -n" in the global zone at the same time that it is running in a non-global zone, can cause the command to fail in one or the other.

If we have puppet running this command, it causes the puppet run to fail for that specific resource.

This patch removes some unnecessary calls to this command, and appropriately updates rspec.