Update validate_data_hash to return modified hash
Puppet::Pops::Lookup::ModuleDataProvider#validate_data_hash is a method
that is supposed to prune a data hash of any keys that are not prefixed
with a given module's name. However prior to this commit it incorrectly
took the data hash, cloned it, operated on the clone, then returned the
unchanged original hash.

This commit updates validate_data_hash to instead return the modified
hash, and updates the warning message generated when a key is pruned to
include the key.